gcp/backend/escritura-bigquery/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.26.0